change plugin name to RpcSecurity
